# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/operators:
    get:
      operationId: getOperators
      tags: [admin]
      summary: List operator accounts
      security:
        - OwnerSession: []
      responses:
        "200":
          description: Operator accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Operator"
  /admin/operators/invitations:
    post:
      operationId: createOperatorInvitation
      tags: [admin]
      summary: Invite an operator by verified email
      security:
        - OwnerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorInvitation"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
  /admin/operators/{id}/approve:
    post:
      operationId: approveOperator
      tags: [admin]
      summary: Approve an invited operator
      security:
        - OwnerSession: []
      parameters:
        - $ref: "#/components/parameters/OperatorId"
      responses:
        "200":
          description: Operator approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
  /admin/operators/{id}/suspend:
    post:
      operationId: suspendOperator
      tags: [admin]
      summary: Suspend an operator
      security:
        - OwnerSession: []
      parameters:
        - $ref: "#/components/parameters/OperatorId"
      responses:
        "200":
          description: Operator suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
  /admin/operators/{id}:
    delete:
      operationId: removeOperator
      tags: [admin]
      summary: Remove an operator
      security:
        - OwnerSession: []
      parameters:
        - $ref: "#/components/parameters/OperatorId"
      responses:
        "204":
          description: Operator removed
  /admin/operator-audit:
    get:
      operationId: getOperatorAuditLogs
      tags: [admin]
      summary: List staff authorization audit records
      security:
        - OwnerSession: []
      responses:
        "200":
          description: Staff authorization audit records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperatorAuditLog"
...
components:
  parameters:
...
    OperatorId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    CustomerSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk customer session. Order ownership is derived from this authenticated session on the server.
    OperatorSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk operator session. The authenticated user's verified email must match an active server-side operator account.
    OwnerSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk owner session. The authenticated user must match the single active owner account in the server-side operator registry.
  schemas:
...
    Operator:
      type: object
      required: [id, email, role, status, linkedToClerk, createdAt, updatedAt]
      properties:
        id: { type: string }
        email: { type: string }
        role: { type: string, enum: [owner, operator] }
        status: { type: string, enum: [invited, active, suspended, removed] }
        linkedToClerk: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    OperatorInvitation:
      type: object
      required: [email]
      properties:
        email:
          type: string
          maxLength: 320
          pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$'
    OperatorAuditLog:
      type: object
      required: [id, action, createdAt]
      properties:
        id: { type: string }
        action: { type: string }
        targetEmail: { type: string }
        createdAt: { type: string }
...